+++ sshd-core/src/main/java/org/apache/sshd/server/command/CommandFactory.java
@@ -19,7 +19,11 @@
 package org.apache.sshd.server.command;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
+import org.apache.sshd.common.util.GenericUtils;
 import org.apache.sshd.server.channel.ChannelSession;
 
 /**
@@ -30,7 +34,6 @@ import org.apache.sshd.server.channel.ChannelSession;
  */
 @FunctionalInterface
 public interface CommandFactory {
-
     /**
      * Create a command with the given name.
      * If the command is not known, a dummy command should be returned to allow
@@ -43,4 +46,69 @@ public interface CommandFactory {
      * @throws IOException if failed to create the instance
      */
     Command createCommand(ChannelSession channel, String command) throws IOException;
+
+    /**
+     * @param command The raw command - ignored if {@code null}/empty
+     * @return The parsed command elements while stripping quoted arguments
+     */
+    static List<String> split(String command) {
+        int len = GenericUtils.length(command);
+        if (len <= 0) {
+            return Collections.emptyList();
+        }
+
+        int curPos = command.indexOf(' ');
+        if (curPos < 0) {
+            return Collections.singletonList(command);
+        }
+
+        int lastPos = 0;
+        List<String> elements = new ArrayList<>();
+        for (curPos = 0; curPos < len; curPos++) {
+            char ch = command.charAt(curPos);
+            // delimited element ?
+            if (((ch == '\'') || (ch == '"'))
+                    // not last character
+                    && (curPos < (len - 1))
+                    // either 1st character or preceded by space
+                    && ((curPos == 0) || (command.charAt(curPos - 1) == ' '))) {
+                // find matching delimiter
+                int nextPos = command.indexOf(ch, curPos + 1);
+                if (nextPos <= curPos) {
+                    continue;   // if not found assume unquoted
+                }
+
+                String elem = command.substring(curPos + 1, nextPos);
+                elements.add(elem);
+
+                curPos = nextPos;
+            } else if (ch != ' ') {
+                continue;
+            } else {
+                if (lastPos < curPos) {
+                    String elem = command.substring(lastPos, curPos);
+                    elements.add(elem);
+                }
+            }
+
+            // skip space and any sequence of them
+            for (curPos++; curPos < len; curPos++) {
+                ch = command.charAt(curPos);
+                if (ch != ' ') {
+                    break;
+                }
+            }
+
+            lastPos = curPos;
+            curPos--;   // compensate for loop auto-increment
+        }
+
+        // any trailing element ?
+        if (lastPos < len) {
+            String elem = command.substring(lastPos);
+            elements.add(elem);
+        }
+
+        return elements;
+    }
 }